Resolve big confusion
After reviewing the git commit history, I had previously deployed the NotificationService
successfully through terraform but then reverted the terraform files. This was done
while trying to resolve the problem with git origin.

This commit has revised the terraform to deploy both lambdas, as well as the S3 resource.

The probably still persists where the IAM policy documents and IAM policies are defined in
the wrong place. These should actually be created outside of the `modules` folder. This is because
right now, the current pattern obliges every lambda to set variables for resources they don't use.